+++ src/components/Popover/index.ts
@@ -1,0 +1,3 @@
+import "./Popover";
+export { type CandyPopoverProps } from "./Popover";
+export { PopoverSide } from "./Types";